Fix #107 (NativeFileIO and large files)
Former-commit-id: a28ebf66acc275f0e68b25a081887a4b685eedbf [formerly 3c03202fcf86024f7ce557af71e59e01a7b24505]
Former-commit-id: c596abef975f1a6ecc54d651d98c27d37e930a9a

+++ av/NativeAV.h
@@ -76,7 +76,7 @@ class NativeAVStreamReader : public NativeAVReader, public NativeStreamDelegate
 
         size_t streamRead;
         size_t streamPacketSize;
-        size_t streamSize;
+        off_t streamSize;
         unsigned const char* streamBuffer;
         int error;
 };
